renamed accordion document files
