+++ src/app/components/hris/employees/employee-profile/employee-profile.component.html
@@ -358,9 +358,9 @@
                 (updateBanking)="updateBankingProgress($event)"></app-accordion-banking>
               <app-accordion-documents-starterkit [hidden]="selectedAccordion != 'Documents'" [employeeProfile]="employeeProfile"
                 (updateDocument)="updateDocumentProgress($event)"></app-accordion-documents-starterkit>
-              <app-accordion-documents-additional [hidden]="selectedAccordion != 'Documents'" [employeeProfile]="employeeProfile"
-                (updateDocument)="updateDocumentProgress($event)"></app-accordion-documents-additional>
-              <app-accordion-documents-custom-documents [hidden]="selectedAccordion != 'Documents'" ></app-accordion-documents-custom-documents>
+              <app-accordion-my-documents [hidden]="selectedAccordion != 'Documents'" [employeeProfile]="employeeProfile"
+                (updateDocument)="updateDocumentProgress($event)"></app-accordion-my-documents>
+              <app-accordion-additional-documents [hidden]="selectedAccordion != 'Documents'" ></app-accordion-additional-documents>
 
             </div>
     </div>
